requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pymongo==4.6.1
mysqlclient==2.2.1
python-dotenv==1.0.0
django-cors-headers==4.3.1
pyjwt==2.8.0
bcrypt==4.0.1
requests==2.31.0
mysql-connector-python==8.2.0
requests==2.31.0